🛠👩👨🥊 refactor user api to be more robust 🛠👩👨🥊

+++ app/api/users/delete/route.ts
@@ -1,11 +1,25 @@
-import storage from "@/models/index";
+import storage from "../../../../models/index";
+import { NextResponse } from "next/server";
 
 export async function POST(request: Request) {
-  const body = await request.json();
+  try {
+    const body = await request.json();
 
-  const id = body.id;
+    if (!body.id) {
+      return NextResponse.json("User id required");
+    }
 
-  storage.deleteUser(id);
+    const id = body.id;
+    try {
+      storage.deleteUser(id);
 
-  storage.save();
+      storage.save();
+
+      return NextResponse.json({ message: "User deleted successfully" });
+    } catch (error) {
+      return NextResponse.json({ message: "Failed to delete User" });
+    }
+  } catch (error) {
+    return NextResponse.json("Invalid body");
+  }
 }
